Move isValidWsExportTarget out of utils
