+++ src/libs/utils.test.ts
@@ -1,11 +1,9 @@
-import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/analysis/_testData/testData';
 import {
   condTyped,
   DEFAULT,
   differenceFromDatesInSeconds,
   differenceFromNowInSeconds,
   formatBytes,
-  isValidWsExportTarget,
   textMatch,
 } from 'src/libs/utils';
 
@@ -44,116 +42,6 @@ describe('differenceFromDatesInSeconds', () => {
     expect(differenceFromDatesInSeconds(startDate, threeSecondsLater)).toBe(3);
     expect(differenceFromDatesInSeconds(startDate, oneMinuteLater)).toBe(60);
     expect(differenceFromDatesInSeconds(startDate, twoDaysLater)).toBe(172800);
-  });
-});
-
-describe('isValidWsExportTarget', () => {
-  it('Returns true because source and dest workspaces are the same', () => {
-    // Arrange
-    const sourceWs = {
-      ...defaultGoogleWorkspace,
-      workspace: {
-        ...defaultGoogleWorkspace,
-        authorizationDomain: [{}],
-      },
-    };
-
-    const destWs = {
-      ...defaultGoogleWorkspace,
-      workspace: {
-        ...defaultGoogleWorkspace.workspace,
-        workspaceId: 'test-different-workspace-id',
-        authorizationDomain: [{}],
-      },
-    };
-
-    // Act
-    const result = isValidWsExportTarget(sourceWs, destWs);
-
-    // Assert
-    expect(result).toBe(true);
-  });
-
-  it('Returns false match because source and dest workspaces are the same', () => {
-    // Arrange
-    const sourceWs = defaultGoogleWorkspace;
-    const destWs = defaultGoogleWorkspace;
-
-    // Act
-    const result = isValidWsExportTarget(sourceWs, destWs);
-
-    // Assert
-    expect(result).toBe(false);
-  });
-
-  it('Returns false because AccessLevel does not contain Writer', () => {
-    // Arrange
-    const sourceWs = defaultGoogleWorkspace;
-    const destWs = {
-      ...defaultGoogleWorkspace,
-      accessLevel: 'READER',
-      workspace: {
-        ...defaultGoogleWorkspace.workspace,
-        workspaceId: 'test-different-workspace-id',
-      },
-    };
-
-    // Act
-    const result = isValidWsExportTarget(sourceWs, destWs);
-
-    // Assert
-    expect(result).toBe(false);
-  });
-
-  it('Returns false because source and destination cloud platforms are not the same.', () => {
-    // Arrange
-    const sourceWs = {
-      ...defaultGoogleWorkspace,
-      workspace: {
-        ...defaultGoogleWorkspace.workspace,
-        authorizationDomain: [{}],
-      },
-    };
-
-    const destWs = {
-      ...defaultAzureWorkspace,
-      workspace: {
-        ...defaultAzureWorkspace.workspace,
-        authorizationDomain: [{}],
-      },
-    };
-
-    // Act
-    const result = isValidWsExportTarget(sourceWs, destWs);
-
-    // Assert
-    expect(result).toBe(false);
-  });
-
-  it('Returns false because source and destination cloud platforms are not the same.', () => {
-    // Arrange
-    const sourceWs = {
-      ...defaultGoogleWorkspace,
-      workspace: {
-        ...defaultGoogleWorkspace.workspace,
-        authorizationDomain: [{}],
-      },
-    };
-
-    const destWs = {
-      ...defaultGoogleWorkspace,
-      workspace: {
-        ...defaultGoogleWorkspace.workspace,
-        authorizationDomain: [{ membersGroupName: 'wooo' }],
-        workspaceId: 'test-different-workspace-id',
-      },
-    };
-
-    // Act
-    const result = isValidWsExportTarget(sourceWs, destWs);
-
-    // Assert
-    expect(result).toBe(false);
   });
 });
